moved dcss-related code into Recoverable.hpp/cpp
