+++ src/persist/Recoverable.cpp
@@ -22,3 +22,23 @@ void Recoverable::init_thread(GlobalTestConfig*, LocalTestConfig* ltc){
 void Recoverable::init_thread(int tid){
     _esys->init_thread(tid);
 }
+
+namespace pds{
+
+    void sc_desc_t::try_complete(pds::EpochSys* esys, uint64_t addr){
+        nbptr_t _d = nbptr.load();
+        int ret = 0;
+        if(_d.val!=addr) return;
+        if(in_progress(_d)){
+            if(pds::esys->check_epoch(cas_epoch)){
+                ret = 2;
+                ret |= commit(_d);
+            } else {
+                ret = 4;
+                ret |= abort(_d);
+            }
+        }
+        cleanup(_d);
+    }
+
+}
