+++ src/persist/Recoverable.hpp
@@ -206,4 +206,163 @@ T* TOKEN_CONCAT(set_, n)(int i, t TOKEN_CONCAT(tmp_, n)){\
 }
 
 
+namespace pds{
+
+    template<typename T>
+    void atomic_nbptr_t<T>::store(const T& desired){
+        // this function must be used only when there's no data race
+        nbptr_t r = nbptr.load();
+        nbptr_t new_r(reinterpret_cast<uint64_t>(desired),r.cnt);
+        nbptr.store(new_r);
+    }
+
+#ifdef VISIBLE_READ
+    // implementation of load and cas for visible reads
+
+    template<typename T>
+    nbptr_t atomic_nbptr_t<T>::load(){
+        nbptr_t r;
+        while(true){
+            r = nbptr.load();
+            nbptr_t ret(r.val,r.cnt+1);
+            if(nbptr.compare_exchange_strong(r, ret))
+                return ret;
+        }
+    }
+
+    template<typename T>
+    nbptr_t atomic_nbptr_t<T>::load_verify(){
+        assert(pds::esys->epochs[pds::EpochSys::tid].ui != NULL_EPOCH);
+        nbptr_t r;
+        while(true){
+            r = nbptr.load();
+            if(pds::esys->check_epoch(pds::esys->epochs[pds::EpochSys::tid].ui)){
+                nbptr_t ret(r.val,r.cnt+1);
+                if(nbptr.compare_exchange_strong(r, ret)){
+                    return r;
+                }
+            } else {
+                throw EpochVerifyException();
+            }
+        }
+    }
+
+    template<typename T>
+    bool atomic_nbptr_t<T>::CAS_verify(nbptr_t expected, const T& desired){
+        assert(pds::esys->epochs[pds::EpochSys::tid].ui != NULL_EPOCH);
+        if(pds::esys->check_epoch(pds::esys->epochs[pds::EpochSys::tid].ui)){
+            nbptr_t new_r(reinterpret_cast<uint64_t>(desired),expected.cnt+1);
+            return nbptr.compare_exchange_strong(expected, new_r);
+        } else {
+            return false;
+        }
+    }
+
+    template<typename T>
+    bool atomic_nbptr_t<T>::CAS(nbptr_t expected, const T& desired){
+        nbptr_t new_r(reinterpret_cast<uint64_t>(desired),expected.cnt+1);
+        return nbptr.compare_exchange_strong(expected, new_r);
+    }
+
+#else /* !VISIBLE_READ */
+    /* implementation of load and cas for invisible reads */
+
+    template<typename T>
+    nbptr_t atomic_nbptr_t<T>::load(){
+        nbptr_t r;
+        do { 
+            r = nbptr.load();
+            if(r.is_desc()) {
+                sc_desc_t* D = r.get_desc();
+                D->try_complete(pds::esys, reinterpret_cast<uint64_t>(this));
+            }
+        } while(r.is_desc());
+        return r;
+    }
+
+    template<typename T>
+    nbptr_t atomic_nbptr_t<T>::load_verify(){
+        // invisible read doesn't need to verify epoch even if it's a
+        // linearization point
+        // this saves users from catching EpochVerifyException
+        return load();
+    }
+
+    // extern std::atomic<size_t> abort_cnt;
+    // extern std::atomic<size_t> total_cnt;
+
+    template<typename T>
+    bool atomic_nbptr_t<T>::CAS_verify(nbptr_t expected, const T& desired){
+        assert(pds::esys->epochs[pds::EpochSys::tid].ui != NULL_EPOCH);
+        // total_cnt.fetch_add(1);
+#ifdef USE_TSX
+        unsigned status = _xbegin();
+        if (status == _XBEGIN_STARTED) {
+            nbptr_t r = nbptr.load();
+            if(!r.is_desc()){
+                if( r.cnt!=expected.cnt ||
+                    r.val!=expected.val ||
+                    !pds::esys->check_epoch(pds::esys->epochs[pds::EpochSys::tid].ui)){
+                    _xend();
+                    return false;
+                } else {
+                    nbptr_t new_r (reinterpret_cast<uint64_t>(desired), r.cnt+4);
+                    nbptr.store(new_r);
+                    _xend();
+                    return true;
+                }
+            } else {
+                // we only help complete descriptor, but not retry
+                _xend();
+                r.get_desc()->try_complete(pds::esys, reinterpret_cast<uint64_t>(this));
+                return false;
+            }
+            // execution won't reach here; program should have returned
+            assert(0);
+        }
+#endif
+        // txn fails; fall back routine
+        // abort_cnt.fetch_add(1);
+        nbptr_t r = nbptr.load();
+        if(r.is_desc()){
+            sc_desc_t* D = r.get_desc();
+            D->try_complete(pds::esys, reinterpret_cast<uint64_t>(this));
+            return false;
+        } else {
+            if( r.cnt!=expected.cnt || 
+                r.val!=expected.val) {
+                return false;
+            }
+        }
+        // now r.cnt must be ..00, and r.cnt+1 is ..01, which means "nbptr
+        // contains a descriptor" and "a descriptor is in progress"
+        assert((r.cnt & 3UL) == 0UL);
+        new (&pds::esys->local_descs[pds::EpochSys::tid].ui) sc_desc_t(r.cnt+1, 
+                                    reinterpret_cast<uint64_t>(this), 
+                                    expected.val, 
+                                    reinterpret_cast<uint64_t>(desired), 
+                                    pds::esys->epochs[pds::EpochSys::tid].ui);
+        nbptr_t new_r(reinterpret_cast<uint64_t>(&pds::esys->local_descs[pds::EpochSys::tid].ui), r.cnt+1);
+        if(!nbptr.compare_exchange_strong(r,new_r)){
+            return false;
+        }
+        pds::esys->local_descs[pds::EpochSys::tid].ui.try_complete(pds::esys, reinterpret_cast<uint64_t>(this));
+        if(pds::esys->local_descs[pds::EpochSys::tid].ui.committed()) return true;
+        else return false;
+    }
+
+    template<typename T>
+    bool atomic_nbptr_t<T>::CAS(nbptr_t expected, const T& desired){
+        // CAS doesn't check epoch; just cas ptr to desired, with cnt+=4
+        assert(!expected.is_desc());
+        nbptr_t new_r(reinterpret_cast<uint64_t>(desired), expected.cnt + 4);
+        if(!nbptr.compare_exchange_strong(expected,new_r)){
+            return false;
+        }
+        return true;
+    }
+
+#endif /* !VISIBLE_READ */
+} // namespace pds
+
 #endif
