+++ src/rideables/MontageLfHashTable.hpp
@@ -17,7 +17,6 @@
 #include "RCUTracker.hpp"
 #include "CustomTypes.hpp"
 #include "Recoverable.hpp"
-#include "DCSS.hpp"
 
 template <class K, class V>
 class MontageLfHashTable : public RMap<K,V>, Recoverable{
